Summary: Added Touch support, sounds, and fixed bugs.
Details:
Added placement sounds and music.
The knight now uses it's current position as the start position.
The animation stops if you interrupt it.
Added Touch support.

Known issues:
Placing the target in the top left [1,8] results in a browser freeze.
This might be a problem with the algorithm, and it is currently being worked on.
Temp fix: Don't put the target in the top left.