Fixed top left cell problem.
Details: Fixed top left cell placement causing browser hang. Did so by setting the default shortest known path length to 12.
This is twice as long as the shortest should ever be, but 5 times shorter than it used to be.
Now the algorithm throws away any paths that exceed 12 steps, reducing the number of possible paths to 823,543.
( Please note this is worst case without out checking the validity of the move outside of insuring it's not the parent move, so the real number should be much lower )